+++ src/libtiled/varianttomapconverter.cpp
@@ -783,6 +783,8 @@ TextData VariantToMapConverter::toTextData(const QVariantMap &variant) const
         alignment |= Qt::AlignHCenter;
     else if (hAlignString == QLatin1String("right"))
         alignment |= Qt::AlignRight;
+    else if (hAlignString == QLatin1String("justify"))
+        alignment |= Qt::AlignJustify;
     else
         alignment |= Qt::AlignLeft;